packages/ai_agent/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
Cython==3.0.12
datasets==3.5.1
dill==0.3.8
fire==0.7.0
frozenlist==1.6.0
idna==3.10
imageio==2.37.0
jieba==0.42.1
joblib==1.5.2
lazy_loader==0.4
lmdb==1.6.2
loguru==0.7.3
lxml==5.4.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numpy==1.26.4
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
paddle-bfloat==0.1.7
pandas==2.3.2
pillow==11.3.0
propcache==0.3.1
pyahocorasick==2.1.0
pyarrow==20.0.0
pyclipper==1.3.0.post6
pydantic==2.11.4
pydantic_core==2.33.2
pypinyin==0.54.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
pytz==2025.2
RapidFuzz==3.13.0
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
shapely==2.1.0
simsimd==6.2.1
six==1.17.0
soupsieve==2.7
stringzilla==3.12.5
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.3.30
torch @ file:///D:/pytorch/torch-1.11.0%2Bcu115-cp310-cp310-win_amd64.whl#sha256=6c66502d4e30464abd8ede9b00ef85ac7eaf569bdf53663375a0ed3f49c4f1e5
torchaudio @ file:///D:/pytorch/torchaudio-0.11.0%2Bcu115-cp310-cp310-win_amd64.whl#sha256=2e7528c87647161c053215a75cae02fd6333e0f1a94825019668cd60446cb746
torchvision @ file:///D:/pytorch/torchvision-0.12.0%2Bcu115-cp310-cp310-win_amd64.whl#sha256=8b7cdcda36bd5af9c0a1c934a680a48ddfe67df7b2710d2b7f61ef06850662dd
typing-inspection==0.4.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
win32_setctime==1.2.0
xxhash==3.5.0
yarl==1.20.0
supabase>=2.0.0
postgrest>=0.10.0